admin/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-restframework==0.0.1
pytz==2021.3
sqlparse==0.4.2
pip~=21.3.1
psycopg2~=2.9.1
colorama~=0.4.4
PyYAML~=5.4.1
Jinja2~=3.0.1
geoip2~=2.9.0
numpy~=1.19.5
setuptools~=57.0.0
Cython~=0.29.21
djangorestframework~=3.12.2
requests~=2.26.0
Markdown~=3.3.4
Pygments~=2.7.1
pika
django-cors-headers
django-filter~=2.0.0
django-rest-knox